# docs/swagger.yaml
basePath: /
definitions:
...
  models.ProgramaEstudio:
    description: Modelo de Programa de Estudio utilizado en la API y sincronizado
      como categoría padre en Moodle.
    properties:
      descripcion:
        example: Programa de estudios enfocado en el desarrollo de software y sistemas
          de información
        type: string
      id_externo:
        example: PROG-ISC-2025
        type: string
      id_moodle:
        example: 9012
        type: integer
      nombre:
        example: Ingeniería en Sistemas Computacionales
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API para gestión de Programa de Estudio, Cuatrimestres, Asignaturas,
    Usuarios y Grupos con sincronización a Moodle.
  title: Control Escolar API
  version: "2.0"
paths:
...
  /programa-estudio/sync/{id}:
    post:
      description: Sincroniza un programa de estudio local con Moodle como categoría
        padre
      parameters:
      - description: ID del programa de estudio a sincronizar
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: Sincronización iniciada correctamente
          schema:
            type: string
        "400":
          description: ID inválido
          schema:
            type: string
        "500":
          description: Error durante la sincronización
          schema:
            type: string
      summary: Sincronizar programa de estudio con Moodle
      tags:
      - ProgramaEstudio
  /programa_estudio:
    get:
      description: Recupera la lista completa de programas de estudio
      produces:
      - application/json
      responses:
        "200":
          description: Lista de programas de estudio
          schema:
            items:
              $ref: '#/definitions/models.ProgramaEstudio'
            type: array
        "500":
          description: Error al obtener programas de estudio
          schema:
            type: string
      summary: Obtener todos los programas de estudio
      tags:
      - ProgramaEstudio
    post:
      consumes:
      - application/json
      description: Crea un nuevo programa de estudio en la base de datos local
      parameters:
      - description: Datos del programa de estudio a crear
        in: body
        name: programa_estudio
        required: true
        schema:
          $ref: '#/definitions/models.ProgramaEstudio'
      produces:
      - application/json
      responses:
        "201":
          description: Programa de estudio creado exitosamente
          schema:
            $ref: '#/definitions/models.ProgramaEstudio'
        "400":
          description: Error en los datos de entrada o campos obligatorios faltantes
          schema:
            type: string
        "500":
          description: Error interno del servidor al crear el programa de estudio
          schema:
            type: string
      summary: Crear un nuevo programa de estudio
      tags:
      - ProgramaEstudio
  /programa_estudio/{id}:
    delete:
      description: Elimina un programa de estudio de la base de datos local
      parameters:
      - description: ID del programa de estudio a eliminar
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Programa de estudio eliminado exitosamente
        "400":
          description: ID inválido
          schema:
            type: string
        "500":
          description: Error al eliminar el programa de estudio
          schema:
            type: string
      summary: Eliminar programa de estudio
      tags:
      - ProgramaEstudio
    get:
      description: Recupera un programa de estudio específico mediante su ID
      parameters:
      - description: ID del programa de estudio
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Programa de estudio encontrado
          schema:
            $ref: '#/definitions/models.ProgramaEstudio'
        "400":
          description: ID inválido
          schema:
            type: string
        "404":
          description: Programa de estudio no encontrado
          schema:
            type: string
      summary: Obtener programa de estudio por ID
      tags:
      - ProgramaEstudio
    put:
      consumes:
      - application/json
      description: Actualiza un programa de estudio existente en la base de datos
        local
      parameters:
      - description: ID del programa de estudio a actualizar
        in: path
        name: id
        required: true
        type: integer
      - description: Datos actualizados del programa de estudio
        in: body
        name: programa_estudio
        required: true
        schema:
          $ref: '#/definitions/models.ProgramaEstudio'
      produces:
      - application/json
      responses:
        "200":
          description: Programa de estudio actualizado exitosamente
          schema:
            $ref: '#/definitions/models.ProgramaEstudio'
        "400":
          description: ID inválido o error en los datos de entrada
          schema:
            type: string
        "500":
          description: Error al actualizar el programa de estudio
          schema:
            type: string
      summary: Actualizar programa de estudio
      tags:
      - ProgramaEstudio
...
swagger: "2.0"